Create Power BI DAX Calendar Code
Learn how you can create Power BI DAX Calendar. You’ll need to create a calendar table in Power BI desktop in a variety of situations. This could be as basic as making a master date table or as sophisticated as making a monthly or weekly index number that is related to a date. There are two approaches for generating a date table for Power BI DAX Calendar. The first way is to generate the table directly in Power BI Desktop, while the second method is to load the date table from the data source. In this video I will show you how to do that in 10 seconds.